requirements.txt
aiogram  # Убедитесь, что версия соответствует используемой вами
google-generativeai  # Убедитесь, что версия соответствует используемой вами
aiohttp